overlay|refactor: Rename buildEntries to entry-outcomes
Renames overlay's `modes/buildEntries.ts` to `modes/entry-outcomes.ts`, a name describing the module's contents rather than a function it does not export.

The module exports the `partitionStatus` and `countOutcome` peers alongside the `OutcomeMap` and `PartitionedStatus` types, so it has no single main export and takes a kebab-case name under the file-naming rule. Its three importers, `modes/create.ts`, `modes/force.ts`, and `modes/verify.ts`, resolve the new path. The module's contents are unchanged, and `src/index.ts` does not reach it, so the published surface is untouched.
