+++ packages/overlay/src/modes/create.ts
@@ -4,7 +4,7 @@ import { parseStatus } from '../chezmoi/parseStatus.ts';
 import { readStatus } from '../chezmoi/readStatus.ts';
 import type { ChezmoiContext } from '../chezmoi/runChezmoi.ts';
 import { runChezmoiStreamed } from '../chezmoi/runChezmoi.ts';
-import { countOutcome, partitionStatus } from './buildEntries.ts';
+import { countOutcome, partitionStatus } from './entry-outcomes.ts';
 import type { OverlayResult } from './types.ts';
 
 /**
